Refactor track routes into modular structure
Split track routes into separate modules for better organization:
- Create core, media, sharing, and middleware route modules
- Extract authentication middleware to separate file
- Improve code separation and maintainability
